api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-dotenv
requests
apscheduler
